fix(sideMenu): fix disabled menu links
The fix for another issue had too general of a CSS selector, which
ultimately prevented clicks in the left menu when it had a scrollable
area. This fix narrows in the selector so it only prevents
pointer-events in the menu-content’s content area (still allowing the
menu to close by tapping the content, but not allowing the content to
be interacted with, like swiping a list item). Closes #1388