revert "Add all angular dependencies to update-angular.sh"
Addresses #3274.

These dependencies should not be in bower.json because they are not used
in Ionic.

Many people's build processes pull in all of these and
automatically bundle them into the browser.  If they do that, it will
break with the addition of `scenario` and `mocks`. Additionally, we
will add unneeded overhead with `aria`, `messages` and `resources`.