feat(confg): create a method to access the global app injector which contains references the bootstr
This feature simply sets a reference to the injector on the IonicApp class, so it can be referenced
by other components. This is sometimes necessary when injecting providers that depend on other
providers. This issue is discussed here
https://github.com/angular/angular/issues/4112#issuecomment-139381970, and Brandon Roberts' solution
of an appInjector() method has been used to solve a variety of dependency injection conflicts.
Unfortunately, it requires access to Angular's bootstrap() method, which Ionic handles in the @App
decorator. This fix will create a reference to the appInjector(), so it can be references from
within Ionic components.

closes #5973