perf(item): reorder is only added to the DOM if needed
I have measured the performance impact of this change, since we use the push change detector strategy, the *ngIf is only evaluated once.

Items wrapped around an element with the ListReorder directive will receive a hidden `<ion-reorder>` in their DOM, but items that are not wrapped (i.e. they CAN NOT be reordered) will not even have the `<ion-reorder>` element in their DOM.

fixes #9065